refactor: update firebase with firebaseconfigs of root
